requirements.txt
Flask==3.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0

